% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Area of a circle}
\usage{
area(x)
}
\arguments{
\item{x}{The radius of the intended circle}
}
\value{
The output from \code{\link{print}}
}
\description{
Area of a circle
}
\examples{
area("2")
}
