% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{argpartsort}
\alias{argpartsort}
\title{Returns the indices of the n-smallest values in a vector, or in each column of a matrix}
\usage{
argpartsort(dat, n)
}
\arguments{
\item{dat}{A vector or a matrix}

\item{n}{the number of indices to return}
}
\value{
a vector of integers (for a vector input) or a matrix (for a matrix input)
}
\description{
Returns the indices of the n-smallest values in a vector, or in each column of a matrix
}
\examples{
# input data is a vector
vec <- c(20, 10, 30, 60, 50, 40)
argpartsort(vec, 3)

# input data is a matrix
mat <- matrix(data = c(20, 10, 30, 60, 50, 40, 40, 20, 10, 30, 60, 50), ncol = 2)
argpartsort(mat, 2)
}
