% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BehavioralDataAnalysis-package.R
\docType{package}
\name{BehavioralDataAnalysis}
\alias{BehavioralDataAnalysis}
\alias{BehavioralDataAnalysis-package}
\title{BehavioralDataAnalysis: Bootstrap and Sampling Functions For Behavioral Data Analysis}
\description{
Bootstrap and Sampling Functions For Behavioral Data Analysis
}
\author{
Florent <florent.buisson.oreilly@maskedmails.com>
}
\keyword{internal}
