% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci}
\alias{boot_ci}
\title{Creates a Bootstrap confidence interval for the output of a function}
\usage{
boot_ci(df, fct, B = 100, conf.level = 0.9, cores = 2)
}
\arguments{
\item{df}{data to use}

\item{fct}{function or linear regression formula to run on the data}

\item{B}{number of Bootstrap sampling iterations to run}

\item{conf.level}{confidence interval percentage to use}

\item{cores}{the number of cores to use for parallel processing}
}
\value{
if the input is a function: a vector with two values, the lower and upper bounds of the confidence interval; otherwise, if the input is a linear regression formula, a matrix where each row contains the lower and upper bounds of the confidence interval for the corresponding regression coefficient
}
\description{
Creates a Bootstrap confidence interval for the output of a function
}
\examples{
options(mc.cores = 2)
df1 <- data.frame(
id = 1:6,
x = c(1, 1.5, 5, 5.5, 10, 10.5),
y = c(1, 1.5, 5, 5.5, 10, 10.5))
boot_ci(df1, function(df) mean(df$x), cores = 2)
}
