% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci_fast_linear.R
\name{boot_ci_fast_linear}
\alias{boot_ci_fast_linear}
\title{Creates a Bootstrap confidence interval for the output of a linear regression,
using Rfast}
\usage{
boot_ci_fast_linear(df, formula, B = 100, conf.level = 0.9, cores = 2)
}
\arguments{
\item{df}{data to use}

\item{formula}{linear regression formula to run on the data}

\item{B}{number of Bootstrap sampling iterations to run}

\item{conf.level}{confidence interval percentage to use}

\item{cores}{the number of cores to use for parallel processing}
}
\value{
a matrix where each row contains the lower and upper bounds of the confidence interval for the corresponding regression coefficient
}
\description{
Creates a Bootstrap confidence interval for the output of a linear regression,
using Rfast
}
