% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_linearity.R
\name{check_linearity}
\alias{check_linearity}
\title{Check Linearity}
\usage{
check_linearity(col1, col2)
}
\arguments{
\item{col1}{numerical vector}

\item{col2}{numerical vector}
}
\value{
the estimated correlation coefficient
}
\description{
Checks the assumption of linearity by calculating the Spearman
correlation coefficient between two quantitative variables.
If your correlation coefficient is close to |1|, then that
indicates a strong linear relationship.
}
\examples{
check_linearity(data$height, data$weight)

}
