% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{Check Normality}
\usage{
check_normality(model)
}
\arguments{
\item{model}{linear model}
}
\value{
Quantile-Quantile (Q-Q) plot.
}
\description{
Checks if the true errors of the model satisfy the assumption
of normality. Data points should ideally fall along the
dotted line.
}
\examples{
linear_model <- lm(height ~ weight, data = data)
check_normality(linear_model)

}
