% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variance.R
\name{check_variance}
\alias{check_variance}
\title{Check Variance}
\usage{
check_variance(model)
}
\arguments{
\item{model}{linear model}
}
\value{
Residuals vs fitted values scatterplot
}
\description{
Checks the assumption of equal variance by a scatterplot
of residual vs fitted values.
}
\examples{
check_variance(linear_model)

}
