% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_fisher_p.R
\name{chi_fisher_p}
\alias{chi_fisher_p}
\title{Function which calculates p-value via Chi-square or Fisher exact test.}
\usage{
chi_fisher_p(tbl, var, treatment)
}
\arguments{
\item{tbl}{(\code{tbl}) Dataframe that has variable and treatment columns of interest}

\item{var}{(\code{character}) Name of variable column}

\item{treatment}{(\code{character}) Name of treatment column}
}
\value{
(\code{numeric}) p-value
}
\description{
Function which calculates p-value via Chi-square or Fisher exact test.
}
\examples{

chi_fisher_p(treatment, "outcome", "treatment")
chi_fisher_p(treatment, "gender", "treatment")

}
