% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean Column Names}
\usage{
clean_names(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame with cleaned column names.
}
\description{
Converts column names to lowercase and replaces non-alphanumeric characters.
}
\examples{
clean_names(data.frame("First Name" = 1, "Value-2" = 3))
}
