% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks.R
\name{Cooks}
\alias{Cooks}
\title{Cooks}
\usage{
Cooks(mod)
}
\arguments{
\item{mod}{the fitted linear regression}
}
\value{
a vector of each observation's Cook's distance
}
\description{
Calculate the Cook's distance, a summary of subject i' influence on beta estimates overall. Use this function with `linear_model`.
}
\examples{
mod <- linear_model(Depression ~ Fatalism + R_E + Sex, Patients)
Cooks(mod)

}
