% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Create Model}
\usage{
create_model(design, data)
}
\arguments{
\item{design}{linear model design}

\item{data}{data frame containing the variables in the model design}
}
\value{
a linear model that can be used to check other regression
assumptions, in statistical tests, etc.
}
\description{
This function creates the linear model
}
\examples{
my_model <- create_model(height ~ weight, data = data)

}
