% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{debug_outliers}
\alias{debug_outliers}
\title{Detect Outliers in Numeric Data}
\usage{
debug_outliers(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list with outliers and thresholds.
}
\description{
Identifies outliers using the 1.5 * IQR rule.
}
\examples{
debug_outliers(c(1, 2, 100, 3, 2, 150))
}
