% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffit.R
\name{dffit}
\alias{dffit}
\title{dffit}
\usage{
dffit(mod)
}
\arguments{
\item{mod}{the fitted linear regression}
}
\value{
a vector of DFFIT_i for each observation.
}
\description{
This function is used to calculate the standardized difference in fits. Please use it with function `linear_model`.
}
\examples{
mod <- linear_model(Depression ~ Fatalism + R_E + Sex, Patients)
dffit(mod)

}
