% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data into R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string specifying the name of the file containing the data.}
}
\value{
This function returns the data as a dataframe object.
}
\description{
This is a simple function that will read data from the
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
into R as a dataframe object if the file containing the data exists in the working directory.
This function requires the "readr" package to be installed.d
}
