% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create dataframes of FARS data for desired years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector containing the years of FARS data the user wishes to analyze.}
}
\value{
Returns a dataframe.
}
\description{
Function that returns dataframes of FARS data for years specified by the user. Returns an error if
there is no FARS data for a specified year.
}
\examples{
fars_read_years(c(2012,2013,2014))

}
