% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes number of accidents by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years of data the user wishes to include in the summary.}
}
\value{
Returns a key-value table displaying the number of accidents each month
of the years in the vector that was passed as an argument to the function.
}
\description{
A function that takes a vector of years as input and returns a table that displays
the number of accidents by year and month. This function requires the "dplyr" package
to be installed.
}
\examples{
fars_summarize_years(c(2012,2013,2014))

}
