% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_Script.R
\name{G.Slice}
\alias{G.Slice}
\title{G.Slice}
\usage{
G.Slice(df, fix.c.range1, fix.c.range2, init.pos, jump, base = "Slice")
}
\arguments{
\item{df}{A data frame}

\item{fix.c.range1}{Numeric parameter, integer, to specify the start column to keep}

\item{fix.c.range2}{Numeric parameter, integer, to specify the final column to keep}

\item{init.pos}{Numeric parameter, integer, which determines from which column the slicing starts}

\item{jump}{Numeric parameter, integer, to specify the length of each slice}

\item{base}{Character parameter to name the slices, by default "Slice"}
}
\value{
Slices of the original data frame keeping some specific columns
}
\description{
G.Slice creates slices from an original data frame, while keeps a specified number of
columns, based on how many columns it has and writes them as '.csv' documents.
}
\examples{
G.Slice(df = MS, fix.c.range1 = 1, fix.c.range2 = 5, init.pos = 6, jump = 10)
}
