% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLH.R
\name{GLH}
\alias{GLH}
\title{GLH}
\usage{
GLH(model, T, c = rep(0, nrow(T)))
}
\arguments{
\item{model}{the fitted linear model results}

\item{T}{matrix giving linear combinations of coefficients by rows}

\item{c}{a vector containing the results of linear combinations of coefficients,defaulting to 0.}
}
\value{
The results of F test including the F statistics and p value
}
\description{
Conduct general linear hypothesis test. Use this function with function \code{linear_model}
}
\examples{
mod <- linear_model(Depression ~ Fatalism + R_E + Sex, Patients)
GLH(mod, diag(4))

}
