% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnPackageDev-package.R
\docType{package}
\name{LearnPackageDev-package}
\alias{LearnPackageDev}
\alias{LearnPackageDev-package}
\title{LearnPackageDev: Helps me learn how to make packages out of R code}
\description{
Nothing much.
}
\author{
\strong{Maintainer}: James Crosbie \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
