% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_string.R
\name{length_string}
\alias{length_string}
\title{Length of the string}
\usage{
length_string(x)
}
\arguments{
\item{x}{The length of the string}
}
\value{
The output from \code{\link{print}}
}
\description{
Length of the string
}
\examples{
length_string("A")
}
