% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{linear_model}
\usage{
linear_model(formula, data)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted}

\item{data}{a data frame containing the variables in the model}
}
\value{
a list containing following components:

\code{formula}: the symbolic description of the model

\code{Coefficients}: the estimated coefficients for each predictor

\code{s.e}: the standard error of coefficient estimates

\code{Cov_beta}: the variance-covariance matrix of estimated coefficients

\code{t_value}: the t statistics of each coefficient

\code{p_value}: the p value of corresponding coefficients' t test

\code{sigma_hat}: the square root of MSE

\code{R2}: R square

\code{F_statistics}: the F statistics for the full model

\code{fitted_values}: the fitted mean values

\code{residuals}: the residuals, observed values mins fitted

\code{X}: the model matrix used

\code{Y}: the response used
}
\description{
Used to fit linear models.
}
\examples{
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
data <- data.frame(cbind(weight, group))
lm.D9 <- linear_model(weight ~ group, data)

}
