% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear Regression}
\usage{
linear_regression(predictor, response)
}
\arguments{
\item{predictor}{Predictor variable}

\item{response}{response variable}
}
\value{
Returns a list of the model coefficients, residuals,
and fitted values.
}
\description{
This function fits a simple linear model from two input vectors.
}
