% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_Script.R
\name{Make.RF}
\alias{Make.RF}
\title{Make.RF}
\usage{
Make.RF(df, mtry.p, var.response)
}
\arguments{
\item{df}{A data frame}

\item{mtry.p}{Numeric parameter, integer, which refers to how many variables are taken into account each iteration of the analysis}

\item{var.response}{Character parameter, is the variable response to settle the formula; a variable from the data frame}
}
\value{
Graph for a random forest model
}
\description{
Make.RF function generates MultiDimensional Scaling plot from a regression formula.
}
\examples{
Make.RF(df = MS[-1], mtry.p = 5, var.response = "Dose")
}
