% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes file name with specified year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Specifies the year of data that is of interest to the user.}
}
\value{
This function returns a character string that can be inputted into the 'fars_read' function
to read in FARS data for available years.
}
\description{
Simple function that creates the name of the file containing FARS data for the year specified.
}
\examples{
make_filename(2012)

}
