% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean0}
\alias{mean0}
\title{Mean_function_with_narm}
\usage{
mean0(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
