% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_Script.R
\name{Mer.Slice}
\alias{Mer.Slice}
\title{Mer.Slice}
\usage{
Mer.Slice(name)
}
\arguments{
\item{name}{Character parameter which is the pattern to look for the files to merge}
}
\value{
A single data frame which contains all the files specified
}
\description{
Mer.Slice function merges two or more data frames into one
}
