% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_Script.R
\name{PC.Graph}
\alias{PC.Graph}
\title{PC.Graph}
\usage{
PC.Graph(df, col1, col2, tag, id)
}
\arguments{
\item{df}{A data frame}

\item{col1}{Numeric parameter, integer, first value of the range of columns to remove}

\item{col2}{Numeric parameter, integer, second value of the range of columns to remove}

\item{tag}{Character parameter which is a variable from the data frame that serves as tag on the plot}

\item{id}{Character parameter which is a variable from the data frame that serves as color on the plot}
}
\value{
A PDF file which contains graphs for each combination of principal components that are needed
}
\description{
PC.Graph applies PCA onto a data frame from which obtains how many principal components (PC)
are needed to cover 97% of variability on the data. PCA.Function is designed for
data frames where one needs to remove a range of columns. Moreover, generates the combination between
those PCs and generates a plot for each combination which are gathered in a .pdf file.
}
\examples{
PC.Graph(df = MS, col1 = 1, col2 = 5, tag = "Sample", id = "Maturation")
}
