% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_Script.R
\name{PCA.Function}
\alias{PCA.Function}
\title{PCA.Function}
\usage{
PCA.Function(df, col1, col2, tag, id, pcx = 1, pcy = 2)
}
\arguments{
\item{df}{A data frame}

\item{col1}{Numeric parameter, integer, first value of the range of columns to remove}

\item{col2}{Numeric parameter, integer, second value of the range of columns to remove}

\item{tag}{Character parameter which is a variable from the data frame that serves as tag on the plot}

\item{id}{Character parameter which is a variable from the data frame that serves as color on the plot}

\item{pcx}{Numeric parameter, integer, to specify which principal component one wants to display on x axis}

\item{pcy}{Numeric parameter, integer, to specify which principal component one wants to display on y axis}
}
\value{
A graphical representation of the loading for the two principal components specified
}
\description{
PCA.Function applies PCA onto a data frame and generates a plot from the Principal
Components of interest. PCA.Function is designed for data frames where one need
to remove a range of columns.
}
\examples{
PCA.Function(df = MS, col1 = 1, col2 = 5, tag = 'Sample', id =  'Maturation')
}
