% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mock.R
\name{predict_mock}
\alias{predict_mock}
\title{predict_mock}
\usage{
predict_mock(mod, newobs)
}
\arguments{
\item{mod}{the fitted linear model results}

\item{newobs}{the new observations, for which you would like to make predictions}
}
\value{
a vector of predict values
}
\description{
Used to get the predictions from the results of a fitted linear model. Please use it with function \code{linear_model}.
}
\examples{
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
data <- data.frame(cbind(weight, group))
lm.D9 <- linear_model(weight ~ group, data)

predict_mock(lm.D9, data.matrix("group" = c(1,3,3,4)))

}
