\name{R_package-package}
\alias{R_package-package}
\alias{R_package}
\docType{package}
\title{
\packageTitle{R_package}
}
\description{
\packageDescription{R_package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{R_package}
\packageIndices{R_package}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{R_package}

Maintainer: \packageMaintainer{R_package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
