% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-squared.R
\name{R_squared}
\alias{R_squared}
\title{R_squared}
\usage{
R_squared(predictor, response)
}
\arguments{
\item{predictor}{Predictor variable}

\item{response}{response variable}
}
\value{
Returns multiple R-squared
}
\description{
This function calculates the r-squared value for a simple linear
regression model..
}
