% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\title{residuals}
\usage{
residuals(mod)
}
\arguments{
\item{mod}{the fitted linear regression}
}
\value{
a table containing four types of residuals
}
\description{
Calculate four types of residuals for a fitted linear regression. Use it with linear_model function.
}
\examples{
mod.davis <- linear_model(weight ~ repwt, data=Davis)
residuals(mod.davis)

}
