% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topten.R
\name{topten}
\alias{topten}
\title{Building a model with Top Top features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n x p matrix of n observations
and p predictors}

\item{y}{a vector of length n representinf the response}
}
\value{
a vector of coefficients from the final fitted model with Top Ten Features
}
\description{
This function develops a prediction algorithm
based on a Top Ten features in 'x' that are most
predictive of the target 'y'
}
\details{
This function runs a univariate regression of y
on each predictor in x and calculates the 'p' value
indicating the significance of association. The final
set of 10 predictors is taken from the 10 smallest p-value
}
\seealso{
\code{lm}
}
\author{
Eyvaz Najafli
}
