% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment}
\alias{treatment}
\title{Outcomes of 100 patients by old and new treatments}
\format{
A data frame with 100 rows and 3 variables:
\itemize{
\item \emph{treatment}: treatment, old or new
\item \emph{treatment}: gender, male or female
\item \emph{outcome}: outcome, failure or success
}
}
\usage{
treatment
}
\description{
A dataset containing the genders and outcomes of two
treatment groups of 100 patients.
}
\keyword{datasets}
