% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BehavioralDataAnalysis-package.R
\docType{package}
\name{BehavioralDataAnalysis}
\alias{BehavioralDataAnalysis-package}
\alias{BehavioralDataAnalysis}
\title{BehavioralDataAnalysis: Bootstrap and Sampling Functions For Behavioral Data Analysis}
\description{
Bootstrap and Sampling Functions For Behavioral Data Analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BehavioralDataAnalysis/R_package}
  \item Report bugs at \url{https://github.com/BehavioralDataAnalysis/R_package/issues}
}

}
\author{
Florent <florent.buisson.oreilly@maskedmails.com>
}
\keyword{internal}
